second commit

+++ src/GrowupFrontendBundle/Controller/IdeaController.php
@@ -12,7 +12,9 @@ use GrowupFrontendBundle\Form\IdeaEditType;
 use GrowupFrontendBundle\Form\IdeaType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
-
+use Symfony\Component\DependencyInjection\Tests\Compiler\IInterface;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use GrowupFrontendBundle\Entity\Idea;
 use GrowupFrontendBundle\Entity\User;
 use GrowupFrontendBundle\Form;
@@ -21,6 +23,7 @@ use Symfony\Component\Form\FormBuilder;
 use Symfony\Component\HttpFoundation\Request;
 use GrowupFrontendBundle\Repository\IdeaRepository;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\HttpFoundation\Response;
 
 class IdeaController extends Controller
 {
@@ -28,14 +31,15 @@ class IdeaController extends Controller
      * @Route("/profile/new", name="new_idea")
      */
 
-    public function newIdeaAction(Request $request){
+    public function newIdeaAction(Request $request)
+    {
 
         $idea = new Idea();
         $form = $this->get('form.factory')->create(IdeaType::class, $idea);
 
         $form->handleRequest($request);
         if ($form->isValid()) {
-            $user=$this->getUser();
+            $user = $this->getUser();
             $idea->setCandidate($user);
             //dump($form->getData());die();
             $em = $this->getDoctrine()->getManager();
@@ -53,17 +57,25 @@ class IdeaController extends Controller
      * Deletes a Idea entity.
      *
      * @Route("/profile/{id}", name="idea_delete")
-     * @Method("GET")
+     * @Method("POST")
      */
     public function deleteAction(Request $request, Idea $idea)
     {
+
+        if ($request->isXmlHttpRequest()) {
+
             $em = $this->getDoctrine()->getManager();
             $em->remove($idea);
             $em->flush();
+            $this->get('session')->getFlashBag()->add('success', 'idée supprimée');
+            $response = new JsonResponse();
 
+            return $response;
+        }
 
         return $this->redirectToRoute('candidate_profile');
     }
+
     /**
      * Displays a form to edit an existing Idea entity.
      *
@@ -72,7 +84,6 @@ class IdeaController extends Controller
      */
     public function editAction(Request $request, $id)
     {
-
 
         $em = $this->getDoctrine()->getManager();
 
@@ -85,35 +96,125 @@ class IdeaController extends Controller
         $form = $this->get('form.factory')->create(IdeaEditType::class, $Idea);
 
         if ($request->isMethod('POST') && $form->handleRequest($request)->isValid()) {
+
             // Inutile de persister ici, Doctrine connait déjà notre annonce
             $em->flush();
 
-            $request->getSession()->getFlashBag()->add('notice', 'Annonce bien modifiée.');
-
+            return new Response('News updated successfully');
             //return $this->redirectToRoute('oc_platform_view', array('id' => $Idea->getId()));
         }
 
-        return $this->render('GrowupFrontendBundle:Candidate:_edit-idea.html.twig', array(
-            'idea' => $Idea,
-            'form'   => $form->createView(),
-        ));
+        return $this->render(
+            'GrowupFrontendBundle:Candidate:_edit-idea.html.twig',
+            array(
+                'idea' => $Idea,
+                'form' => $form->createView(),
+            )
+        );
+
+        /*$em = $this->getDoctrine()->getManager();
+        $form = $this->createForm(IdeaEditType::class, $idea);
+
+        if ($request->getMethod() == 'POST') {
+
+            $form->handleRequest($request);
+            if ($form->isValid()) {
+                $em->persist($idea);
+                $this->get('session')->getFlashBag()->add('success', 'Informations modifiées');
+
+            }
+
+        }
+
+        //return $this->redirectToRoute('candidate_profile', array('id' => $idea->getId()));
+
+        return $this->render(
+            'GrowupFrontendBundle:Candidate:_edit-idea.html.twig',
+            array(
+                'idea' => $idea,
+                'form' => $form->createView(),
+            )
+        );*/
     }
+
 
     /**
      *
      * @Route("/{id}/show", name="idea_show")
      */
-    public function showAction($id){
+    public function showAction($id)
+    {
 
         $em = $this->getDoctrine()->getManager();
 
-        $ideas = $em->getRepository('GrowupFrontendBundle:Idea')->findOneBy(array('id'=>$id));
-        return $this->render('GrowupFrontendBundle:Candidate:show-idea.html.twig', array(
-            'idea' => $ideas,
-        ));
+        $ideas = $em->getRepository('GrowupFrontendBundle:Idea')->findOneBy(array('id' => $id));
+
+        return $this->render(
+            'GrowupFrontendBundle:Candidate:show-idea.html.twig',
+            array(
+                'idea' => $ideas,
+            )
+        );
     }
 
+
+    /**
+     *
+     * @Route("ajax/{id}/edit", name="idea_edit_ajax")
+     */
+    public function modifiertAjaxAction(Request $request, Idea $idea)
+    {
+        $form = $this->createForm(IdeaEditType::class, $idea);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted()) {
+
+            if ($request->isXmlHttpRequest()) {
+
+                if ($form->isValid()) {
+                    $em = $this->getDoctrine()->getManager();
+                    $em->persist($idea);
+                    $em->flush();
+
+                    $response = new JsonResponse();
+
+                    return $response;
+
+                } else {
+                    $errors = $this->get('app.form.error.messages')->getErrorMessages($form);
+
+                    $view = $this->renderView(
+                        '@GrowupFrontend/erreurs-form.html.twig',
+                        array(
+                            'data' => $errors,
+                            //'form' => $form->createView(),
+                        )
+                    );
+                    $response = new JsonResponse(
+                        array(
+                            'view' => $view,
+                            'result' => 0,
+                            'message' => 'Formulaire non valide',
+                        )
+                    );
+
+                    return $response;
+                }
+            }
+
+        }
+
+        return $this->render(
+            '@GrowupFrontend/Candidate/edit-ajax.html.twig',
+            array(
+                'form' => $form->createView(),
+                'idea' => $idea,
+            )
+        );
     }
+
+
+}
